1.增加按钮对象 2.增加获取当前光标处RectItem的方法 3.增加FPropertys属性，可自定义存放病历全局相关信息 4.修正保存和读取时Script和FloatCount顺序反了的问题 5.使用树结构来处理数据组的交互 6.修正剪切区域后没有还原原来区域造成后面的绘制都看不到的问题 7.修正不是在开始位置插入内容时，第一个Item是空行且不能和前面合并时引起文本Item为空的异常

+++ Source/Win32.GDI.cs
@@ -804,7 +804,7 @@ namespace HC.Win32
 		[DllImport("gdi32")] public static extern int CreatePolyPolygonRgn(ref POINT lpPoint, ref int lpPolyCounts, int nCount, int nPolyFillMode);
 		[DllImport("gdi32")] public static extern IntPtr CreatePolygonRgn(POINT[] lpPoint, int nCount, int nPolyFillMode);
 		[DllImport("gdi32")] public static extern IntPtr CreateRectRgn(int X1, int Y1, int X2, int Y2);
-		[DllImport("gdi32")] public static extern int CreateRectRgnIndirect(ref RECT lpRect);
+		[DllImport("gdi32")] public static extern IntPtr CreateRectRgnIndirect(ref RECT lpRect);
 		[DllImport("gdi32")] public static extern int CreateRoundRectRgn(int X1, int Y1, int X2, int Y2, int X3, int Y3);
 		[DllImport("gdi32")] public static extern int CreateScalableFontResource(int fHidden, string lpszResourceFile, string lpszFontFile, string lpszCurrentPath);
 		[DllImport("gdi32")] public static extern int CreateSolidBrush(int crColor);
